+++ src/agent/adu_core_interface/src/adu_core_interface.c
@@ -543,7 +543,7 @@ JSON_Value* GetReportingJsonValue(
 
     JSON_Value* rootValue = json_value_init_object();
     JSON_Object* rootObject = json_value_get_object(rootValue);
-    int stepsCount = workflow_get_children_count(handle);
+    size_t stepsCount = workflow_get_children_count(handle);
 
     //
     // Prepare 'lastInstallResult', 'stepResults' data.
@@ -692,7 +692,7 @@ JSON_Value* GetReportingJsonValue(
     if (updateState != ADUCITF_State_DownloadStarted)
     {
         stepsCount = workflow_get_children_count(handle);
-        for (int i = 0; i < stepsCount; i++)
+        for (size_t i = 0; i < stepsCount; i++)
         {
             ADUC_WorkflowHandle childHandle = workflow_get_child(handle, i);
             ADUC_Result childResult;
