+++ src/extensions/step_handlers/swupdate_handler_v2/tests/swupdate_handler_v2_ut.cpp
@@ -138,7 +138,7 @@ TEST_CASE("SWUpdate Prepare Arguments Test")
     result = PrepareStepsWorkflowDataObject(handle);
     CHECK(result.ResultCode != 0);
 
-    int childCount = workflow_get_children_count(handle);
+    size_t childCount = workflow_get_children_count(handle);
     CHECK(childCount == 1);
 
     ADUC_WorkflowHandle stepHandle = workflow_get_child(handle, 0);
@@ -231,7 +231,7 @@ TEST_CASE("SWUpdate Prepare Arguments Test v2.1")
     result = PrepareStepsWorkflowDataObject(handle);
     CHECK(result.ResultCode != 0);
 
-    int childCount = workflow_get_children_count(handle);
+    size_t childCount = workflow_get_children_count(handle);
     CHECK(childCount == 1);
 
     ADUC_WorkflowHandle stepHandle = workflow_get_child(handle, 0);
